Changed graph from template to long. First parsing.

+++ GraphViewer/connection.cpp
@@ -37,7 +37,7 @@ Connection::Connection(short port) {
 	// Create a socket.
     sock = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
     if (sock == INVALID_SOCKET) {
-        printf("Client: socket() - Error at socket(): %ld\n", WSAGetLastError());
+        printf("Client: socket() - Error at socket(): %ld\n", (long int) WSAGetLastError());
         WSACleanup();
     }
